+++ frontend/src/store/types.ts
@@ -9,17 +9,22 @@ export interface AuthState {
   isAuthenticated: boolean;
 }
 
+export interface ChatState {
+  isExpanded: boolean;
+  isSidebarOpen: boolean;
+  // Add other chat-related state properties as needed
+}
+
 export interface RootState {
   auth: AuthState;
+  chat: ChatState;
   // Add other module states as needed
   // For example:
   // studio: StudioState;
   // theme: ThemeState;
-  // chat: ChatState;
 }
 
 // Add other module state interfaces as needed
 // For example:
 // export interface StudioState { ... }
 // export interface ThemeState { ... }
-// export interface ChatState { ... }
